refactor(nest): simplify module registration logic
