+++ packages/framework/src/servers/nest/nest.constants.ts
@@ -1,3 +1,2 @@
-export const API_PATH = 'API_PATH';
 export const REGISTER_API_PATH = 'REGISTER_API_PATH';
-export { NOVU_OPTIONS } from './nest-base.module';
+export { NOVU_OPTIONS } from './nest.module-definition';
